refactor: handler level code refactor + a few import formating
